+++ platform/user/api/src/main/java/com/digiboy/platform/user/api/UserService.java
@@ -1,0 +1,11 @@
+package com.digiboy.platform.user.api;
+
+import com.digiboy.platform.user.dto.UserDTO;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+public interface UserService {
+    Page<UserDTO> findAll(Pageable pageable);
+
+    UserDTO save(UserDTO userDTO);
+}
