+++ platform/user/api/src/main/java/com/digiboy/platform/user/dto/DataTransferObject.java
@@ -1,0 +1,4 @@
+package com.digiboy.platform.user.dto;
+
+public interface DataTransferObject {
+}
