--- platform/user/api/src/main/java/com/digiboy/platform/users/dto/DataTransferObject.java
@@ -1,4 +1,0 @@
-package com.digiboy.platform.users.dto;
-
-public interface DataTransferObject {
-}
