--- platform/user/api/src/main/java/com/digiboy/platform/users/dto/UserDTO.java
@@ -1,46 +1,0 @@
-package com.digiboy.platform.users.dto;
-
-import java.util.UUID;
-
-public class UserDTO implements DataTransferObject {
-
-    private UUID userId;
-
-    private String username;
-
-    private String password;
-
-    private String email;
-
-    public UUID getUserId() {
-        return userId;
-    }
-
-    public void setUserId(UUID userId) {
-        this.userId = userId;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
