+++ platform/user/repository/src/main/java/com/digiboy/platform/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.digiboy.platform.user.repository;
+
+import com.digiboy.platform.user.to.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface UserRepository extends JpaRepository<User, UUID> {
+    boolean existsByUsername(String username);
+}
