+++ platform/user/repository/src/main/java/com/digiboy/platform/user/to/DomainEntity.java
@@ -1,0 +1,4 @@
+package com.digiboy.platform.user.to;
+
+public interface DomainEntity {
+}
