+++ platform/user/repository/src/main/java/com/digiboy/platform/user/to/Role.java
@@ -1,0 +1,4 @@
+package com.digiboy.platform.user.to;
+
+public class Role {
+}
