--- platform/user/repository/src/main/java/com/digiboy/platform/users/to/DomainEntity.java
@@ -1,4 +1,0 @@
-package com.digiboy.platform.users.to;
-
-public interface DomainEntity {
-}
