--- platform/user/repository/src/main/java/com/digiboy/platform/users/to/EntityBase.java
@@ -1,30 +1,0 @@
-package com.digiboy.platform.users.to;
-
-import javax.persistence.*;
-
-@MappedSuperclass
-public abstract class EntityBase {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "SEQUENCE_GENERATOR")
-    private Long id;
-
-    @Version
-    private Long version;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Long getVersion() {
-        return version;
-    }
-
-    public void setVersion(Long version) {
-        this.version = version;
-    }
-}
