--- platform/user/repository/src/main/java/com/digiboy/platform/users/to/User.java
@@ -1,59 +1,0 @@
-package com.digiboy.platform.users.to;
-
-import javax.persistence.*;
-import java.util.UUID;
-
-@Entity
-@Table(name = "users", uniqueConstraints = {
-        @UniqueConstraint(name = "UQ_USERS_USERNAME", columnNames = {"username"}),
-        @UniqueConstraint(name = "UQ_USERS_EMAIL", columnNames = {"email"}),
-})
-public class User implements DomainEntity {
-
-    @Id
-    @Column(name = "user_id")
-    @GeneratedValue(strategy = GenerationType.AUTO)
-//    @Type(type="uuid-char")
-    private UUID userId;
-
-    @Column(name = "username")
-    private String username;
-
-    @Column(name = "password")
-    private String password;
-
-    @Column(name = "email")
-    private String email;
-
-    public UUID getUserId() {
-        return userId;
-    }
-
-    public void setUserId(UUID userId) {
-        this.userId = userId;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
