+++ platform/user/repository/src/test/java/com/digiboy/platform/user/TestConfiguration.java
@@ -1,0 +1,9 @@
+package com.digiboy.platform.user;
+
+import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+
+@SpringBootConfiguration
+@EnableAutoConfiguration
+public class TestConfiguration {
+}
