--- platform/user/repository/src/test/java/com/digiboy/platform/users/TestConfiguration.java
@@ -1,9 +1,0 @@
-package com.digiboy.platform.users;
-
-import org.springframework.boot.SpringBootConfiguration;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-
-@SpringBootConfiguration
-@EnableAutoConfiguration
-public class TestConfiguration {
-}
