+++ platform/user/service/src/main/java/com/digiboy/platform/user/mapper/UserMapper.java
@@ -1,0 +1,9 @@
+package com.digiboy.platform.user.mapper;
+
+import com.digiboy.platform.user.dto.UserDTO;
+import com.digiboy.platform.user.to.User;
+import org.mapstruct.Mapper;
+
+@Mapper
+public interface UserMapper extends EntityMapper<User, UserDTO> {
+}
