+++ platform/user/service/src/main/java/com/digiboy/platform/user/service/ServiceBase.java
@@ -1,0 +1,4 @@
+package com.digiboy.platform.user.service;
+
+public abstract class ServiceBase {
+}
