--- platform/user/service/src/main/java/com/digiboy/platform/users/mapper/UserMapper.java
@@ -1,9 +1,0 @@
-package com.digiboy.platform.users.mapper;
-
-import com.digiboy.platform.users.dto.UserDTO;
-import com.digiboy.platform.users.to.User;
-import org.mapstruct.Mapper;
-
-@Mapper
-public interface UserMapper extends EntityMapper<User, UserDTO> {
-}
