--- platform/user/service/src/main/java/com/digiboy/platform/users/service/ServiceBase.java
@@ -1,4 +1,0 @@
-package com.digiboy.platform.users.service;
-
-public abstract class ServiceBase {
-}
