+++ platform/user/web/src/main/java/com/digiboy/platform/user/web/validators/UsernameConstraint.java
@@ -1,0 +1,15 @@
+package com.digiboy.platform.user.web.validators;
+
+import javax.validation.Constraint;
+import javax.validation.Payload;
+import java.lang.annotation.*;
+
+@Documented
+@Constraint(validatedBy = UsernameValidator.class)
+@Target({ElementType.METHOD, ElementType.FIELD})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface UsernameConstraint {
+    String message() default "Invalid phone number";
+    Class<?>[] groups() default {};
+    Class<? extends Payload>[] payload() default {};
+}
