+++ platform/user/web/src/main/java/com/digiboy/platform/user/web/validators/UsernameValidator.java
@@ -1,0 +1,18 @@
+package com.digiboy.platform.user.web.validators;
+
+import javax.validation.ConstraintValidator;
+import javax.validation.ConstraintValidatorContext;
+
+public class UsernameValidator implements ConstraintValidator<UsernameConstraint, String> {
+    @Override
+    public void initialize(UsernameConstraint constraintAnnotation) {
+        ConstraintValidator.super.initialize(constraintAnnotation);
+    }
+
+    @Override
+    public boolean isValid(String contactField, ConstraintValidatorContext ctx) {
+//        return contactField != null && contactField.matches("[0-9]+")
+//                && (contactField.length() > 8) && (contactField.length() < 14);
+        return true;
+    }
+}
