--- platform/user/web/src/main/java/com/digiboy/platform/users/web/config/security/SecurityConfiguration.java
@@ -1,14 +1,0 @@
-package com.digiboy.platform.users.web.config.security;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-public class SecurityConfiguration {
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-}
