--- platform/user/web/src/main/java/com/digiboy/platform/users/web/config/security/WebSecurity.java
@@ -1,16 +1,0 @@
-package com.digiboy.platform.users.web.config.security;
-
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-
-@EnableWebSecurity
-public class WebSecurity extends WebSecurityConfigurerAdapter {
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http.csrf().disable();
-        http.authorizeRequests()
-                .antMatchers("/actuator/**").permitAll()
-                .antMatchers("/**").permitAll();
-    }
-}
