+++ platform/user/web/src/test/java/com/digiboy/platform/user/ApplicationTest.java
@@ -1,0 +1,16 @@
+package com.digiboy.platform.user;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.core.env.Environment;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@SpringBootTest
+class ApplicationTest {
+
+    @Test
+    void loadContext() {
+    }
+}
